+++ src/main/java/com/clinica/model/repositories/RolRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolRepository extends JpaRepository<Rol, Byte> {
+public interface RolRepository extends JpaRepository<Rol, Long> {
     // Spring Data JPA ya nos da el método findById() por defecto
 }
